BUG/MINOR: rhttp: fix task_wakeup state
TASK_WOKEN_ANY was incorrectly used as argument to task_wakeup() for
rhttp preconnect task. This value is used as a flag. Replace it by
proper individual values. This is labelled as a bug but it has no known
impact.

This should be backported up to 2.9.

(cherry picked from commit 808daa7cfb46019fff0e0bf4580985568f145c17)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>