BUG/MINOR: http-htx: Support default path during scheme based normalization
As stated in RFC3986, for an absolute-form URI, an empty path should be
normalized to a path of "/". This is part of scheme based normalization
rules. This kind of normalization is already performed for default ports. So
we might as well deal with the case of empty path.

The associated reg-tests was updated accordingly.

This patch should fix the issue #2573. It may be backported as far as 2.4 if
necessary.

(cherry picked from commit 8d2514e0872a597e2f72659924aa67352b78c4d4)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>
(cherry picked from commit eff074a34c78a784c706955e264722fc7cb242b0)
Signed-off-by: Amaury Denoyelle <adenoyelle@haproxy.com>