BUG/MEDIUM: stick-tables: Fix locking of updt_lock when trashing oldest entries
A bug was introduced in commit 80a0ea962d ("BUG/MEDIUM: stick-tables: Fix
race with peers when trashing oldest entries") during the backport
process. In 3.0, there are 2 nested while loops. The update lock is acquired
in the inner loop and released when exiting for it.

But on 2.9, there is only one loop. Thus the update lock is only released at
the end of the function, if it was previously acquired. So let's fix that.

It is a 2.9-specific issue. There is no upstream ID and no backport is
needed, except if commit above is backported.