CI: scripts: reset backported changes on macOS version
The following patch was backported to 2.9 version. This aims at fixing
VTest compilation in some environment

  * 8c4ad96036
    CI: modernize macos matrix

  * c42accdb43
    CI: scripts: fix build of vtest regarding option -C

This first one was not targetted for backport but it was deemed
necessary to apply the second one without conflict. However it also
contains undesirable changes in CI for macOS version support.

This commit fixes this by properly merging changes from the last one
commit alone with conflict changes, without touching to CI
configuration.

This is necessary for 2.9 only as on 2.8 proper backport will be done.