BUG/MEDIUM: cli: fix cli_output_msg() regression
Jens Wahnes reported on the ML [1] that the output of some commands sent
via the socket would be mangled with odd data sometimes.

He mentioned that he only faced the issue while switching from 2.4.26 to
2.4.27 and that he wasn't able to reproduce the issue on 2.8.10. Also,
he mentionned that the bad characters found after the cli output were
likely leftovers from a previous http request.

With all those precious inputs from Jens, finding the offending commit was
pretty easy. It turns out that this regression was introduced by 715674028
("BUG/MEDIUM: cli: fix once for all the problem of missing trailing LFs"),
whose upstream ID is 6219a58.

What happens since that commit is that we strip the end of the string to
remove any existing LF, then we forcefully add one using chunk_iscat(),
and finally we send the message using ci_putblk() using strlen(tmp->area)
as string length. But this is wrong! Indeed, since we strip some bytes off
the original message, it means we may overwrite NULL-byte from original
string, and since chunk_istcat(), unlike chunk_appendf(), doesn't add a
terminating NULL-byte, we cannot safely use strlen() on the chunk's area
anymore. This explains why the function sometimes output garbage data
(data from a previous trash chunk user since the trash isn't memset
between 2 uses) at the end of the expected output.

This bug only affects 2.4 and 2.6: indeed, when backporting ("BUG/MEDIUM:
cli: fix once for all the problem of missing trailing LFs") we didn't
notice that the original commit, which was originally written for the 3.0,
was relying e8ee27b0fd ("MINOR: cli: Use applet API to write output
message") commit (first seen in 2.8) that makes cli_output_msg() func
use applet_putchk() instead of ci_putblk() to push the message.

Using _putchk() to push the buffer instead of using _putblk() + strlen()
combination is indeed safer since the internal buffer size is used instead
of relying on the presence of the terminating NULL byte on the string.

Thus to fix the issue, let's simply switch to using ci_putchk() instead of
ci_putblk()+strlen() combination in cli_output_msg(). Of course, in 2.6,
another solution could have been to backport e8ee27b0fd, but due to API
changes between 2.6 and 2.8, even if the patch seems to easily apply, it
was decided not to backport it and instead provide the same generic fix
for both 2.4 and 2.6 to prevent any side-effect.

This patch has no upstream ID, it was specifically written for the 2.6 and
2.4 where ("BUG/MEDIUM: cli: fix once for all the problem of missing
trailing LFs") was already backported.

It should be backported to 2.4.

[1]: https://www.mail-archive.com/haproxy@formilux.org/msg45072.html